requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.5.1
# email-validator==2.2.2